requirements.txt
flask
flask-cors
python-dotenv
docling
openai
requests
pillow
